[TASK] Hack TF 7 for codeception 5 compatility
Core v12 will switch to codeception 5. testing-framework 7
needs to support both codeception 4 (core v11) and
codeception 5 (core v12) to allow extensions using
acceptance tests when preparing v11 AND v12
compatibility.

codeception 5 has some hard breaking PHP changes, especially
on API class properties. We have no good choice here and
decided to introduce a class_alias switch based on used
core version for, this since extensions need to extend especially
BackendEnvironment to set up acceptance tests based on
codeception.

Releases: 7
