+++ Classes/Core/Acceptance/Helper/Login.php
@@ -1,7 +1,6 @@
 <?php
 
 declare(strict_types=1);
-namespace TYPO3\TestingFramework\Core\Acceptance\Helper;
 
 /*
  * This file is part of the TYPO3 CMS project.
@@ -16,125 +15,23 @@ namespace TYPO3\TestingFramework\Core\Acceptance\Helper;
  * The TYPO3 project - inspiring people to share!
  */
 
-use Codeception\Exception\ConfigurationException;
-use Codeception\Module;
-use Codeception\Module\WebDriver;
-use Codeception\Util\Locator;
+namespace TYPO3\TestingFramework\Core\Acceptance\Helper;
+
+use TYPO3\CMS\Core\Information\Typo3Version;
 
 /**
- * Helper class to log in backend users and load backend.
+ * This is an ugly hack exclusively for testing-framework v7 to allow
+ * both codeception 4 (core v11) and 5 (core v12) at the same time.
+ *
+ * Problem is the codeception API is hard breaking between codeception 4 and 5,
+ * especially due to new type hints on properties that we have to use.
  */
-class Login extends Module
+if (((new Typo3Version())->getMajorVersion() >= 12)) {
+    class_alias(LoginTwelve::class, 'TYPO3\\TestingFramework\\Core\\Acceptance\\Helper\\LoginConditionalParent');
+} else {
+    class_alias(LoginEleven::class, 'TYPO3\\TestingFramework\\Core\\Acceptance\\Helper\\LoginConditionalParent');
+}
+
+class Login extends LoginConditionalParent
 {
-    /**
-     * @var array Filled by .yml config with valid sessions per role
-     */
-    protected $config = [
-        'sessions' => [],
-    ];
-
-    /**
-     * Set a backend user session cookie and load the backend index.php.
-     *
-     * Use this action to change the backend user and avoid switching between users in the backend module
-     * "Backend Users" as this will change the user session ID and make it useless for subsequent calls of this action.
-     *
-     * @param string $role The backend user who should be logged in.
-     * @throws ConfigurationException
-     */
-    public function useExistingSession($role = '')
-    {
-        $webDriver = $this->getWebDriver();
-
-        $newUserSessionId = $this->getUserSessionIdByRole($role);
-
-        $hasSession = $this->_loadSession();
-        if ($hasSession && $newUserSessionId !== '' && $newUserSessionId !== $this->getUserSessionId()) {
-            $this->_deleteSession();
-            $hasSession = false;
-        }
-
-        if (!$hasSession) {
-            $webDriver->amOnPage('/typo3/index.php');
-            $webDriver->waitForElement('body[data-typo3-login-ready]');
-            $this->_createSession($newUserSessionId);
-        }
-
-        // Reload the page to have a logged in backend.
-        $webDriver->amOnPage('/typo3/index.php');
-
-        // Ensure main content frame is fully loaded, otherwise there are load-race-conditions ..
-        $webDriver->waitForElement('iframe[name="list_frame"]');
-        $webDriver->switchToIFrame('list_frame');
-        $webDriver->waitForElement(Locator::firstElement('div.module'));
-        // .. and switch back to main frame.
-        $webDriver->switchToIFrame();
-    }
-
-    /**
-     * @param string $role
-     * @return string
-     * @throws ConfigurationException
-     */
-    protected function getUserSessionIdByRole($role)
-    {
-        if (empty($role)) {
-            return '';
-        }
-
-        if (!isset($this->_getConfig('sessions')[$role])) {
-            throw new ConfigurationException(sprintf(
-                'Backend user session ID cannot be resolved for role "%s": ' .
-                'Set session ID explicitly in configuration of module Login.',
-                $role
-            ), 1627554106);
-        }
-
-        return $this->_getConfig('sessions')[$role];
-    }
-
-    /**
-     * @return bool
-     */
-    public function _loadSession()
-    {
-        return $this->getWebDriver()->loadSessionSnapshot('login');
-    }
-
-    public function _deleteSession()
-    {
-        $webDriver = $this->getWebDriver();
-        $webDriver->resetCookie('be_typo_user');
-        $webDriver->resetCookie('be_lastLoginProvider');
-        $webDriver->deleteSessionSnapshot('login');
-    }
-
-    /**
-     * @param string $userSessionId
-     */
-    public function _createSession($userSessionId)
-    {
-        $webDriver = $this->getWebDriver();
-        $webDriver->setCookie('be_typo_user', $userSessionId);
-        $webDriver->setCookie('be_lastLoginProvider', '1433416747');
-        $webDriver->saveSessionSnapshot('login');
-    }
-
-    /**
-     * @return string
-     */
-    protected function getUserSessionId()
-    {
-        $userSessionId = $this->getWebDriver()->grabCookie('be_typo_user');
-        return $userSessionId ?? '';
-    }
-
-    /**
-     * @return WebDriver
-     * @throws \Codeception\Exception\ModuleException
-     */
-    protected function getWebDriver()
-    {
-        return $this->getModule('WebDriver');
-    }
 }
